+++ src/main/java/com/dementor/domain/mentor/dto/request/MentorRequestDto.java
@@ -1,4 +1,4 @@
-package com.dementor.domain.mentor.dto;
+package com.dementor.domain.mentor.dto.request;
 
 import com.dementor.domain.mentor.entity.MentorEntity;
 import lombok.AllArgsConstructor;
